fixup! feat(core): Add support for hostless components
