+++ goldens/public-api/compiler-cli/error_code.api.md
@@ -68,7 +68,6 @@ export enum ErrorCode {
     HOST_DIRECTIVE_NOT_STANDALONE = 2014,
     HOST_DIRECTIVE_UNDEFINED_BINDING = 2017,
     HOSTLESS_COMPONENT_ANIMATIONS = 2031,
-    HOSTLESS_COMPONENT_HOST_STYLE = 2032,
     HOSTLESS_COMPONENT_SHADOW_DOM = 2030,
     HOSTLESS_COMPONENT_UNSUPPORTED_BINDING = 8030,
     HOSTLESS_COMPONENT_WITH_HOST_BINDINGS = 2029,
