+++ packages/core/src/render3/instructions/element_container.ts
@@ -5,7 +5,16 @@
  * Use of this source code is governed by an MIT-style license that can be
  * found in the LICENSE file at https://angular.dev/license
  */
-import {validateMatchingNode, validateNodeExists} from '../../hydration/error_handling';
+import {
+  invalidSkipHydrationHost,
+  validateMatchingNode,
+  validateNodeExists,
+} from '../../hydration/error_handling';
+import {hasSkipHydrationAttrOnTNode} from '../../hydration/skip_hydration';
+import {markRNodeAsSkippedByHydration} from '../../hydration/utils';
+import {nativeRemoveNode} from '../dom_node_manipulation';
+import {RNode} from '../interfaces/renderer_dom';
+import {isComponentHost} from '../interfaces/type_checks';
 import {locateNextRNode, siblingAfter} from '../../hydration/node_lookup_utils';
 import {
   canHydrateNode,
@@ -23,6 +32,7 @@ import {HEADER_OFFSET, HYDRATION, LView, RENDERER, TVIEW, TView} from '../interf
 import {assertTNodeType} from '../node_assert';
 import {executeContentQueries} from '../queries/query_execution';
 import {
+  enterSkipHydrationBlock,
   getBindingsEnabled,
   getCurrentTNode,
   getLView,
@@ -212,7 +222,7 @@ export function ɵɵdomElementContainer(
   return ɵɵdomElementContainer;
 }
 
-let _locateOrCreateElementContainerNode: typeof locateOrCreateElementContainerNode = (
+export let _locateOrCreateElementContainerNode: typeof locateOrCreateElementContainerNode = (
   tView: TView,
   lView: LView,
   tNode: TNode,
@@ -263,6 +273,26 @@ function locateOrCreateElementContainerNode(
   if (ngDevMode) {
     validateMatchingNode(comment, Node.COMMENT_NODE, null, lView, tNode);
     markRNodeAsClaimedByHydration(comment);
+  }
+
+  if (hydrationInfo && hasSkipHydrationAttrOnTNode(tNode)) {
+    if (isComponentHost(tNode)) {
+      enterSkipHydrationBlock(tNode);
+
+      // Since this isn't hydratable, we need to empty the container's contents
+      // so there's no duplicate content after render.
+      const renderer = lView[RENDERER];
+      let nodeToRemove: RNode | null = currentRNode;
+      while (nodeToRemove && nodeToRemove !== comment) {
+        const next: RNode | null = (nodeToRemove as Node).nextSibling as RNode | null;
+        nativeRemoveNode(renderer, nodeToRemove);
+        nodeToRemove = next;
+      }
+
+      ngDevMode && markRNodeAsSkippedByHydration(comment);
+    } else if (ngDevMode) {
+      throw invalidSkipHydrationHost(comment);
+    }
   }
 
   return comment;
